+++ deepchecks/checks.py
@@ -64,6 +64,7 @@ __all__ = [
     'TrainTestLabelDrift',
     'WholeDatasetDrift',
     'TrainTestPredictionDrift',
+    'MultivariateDrift',
 
     # performance checks
     'TrainTestPerformance',
