+++ deepchecks/tabular/checks/train_test_validation/__init__.py
@@ -24,6 +24,7 @@ from .train_test_feature_drift import TrainTestFeatureDrift
 from .train_test_label_drift import TrainTestLabelDrift
 from .train_test_samples_mix import TrainTestSamplesMix
 from .whole_dataset_drift import WholeDatasetDrift
+from .multivariate_drift import MultivariateDrift
 
 __all__ = [
     'CategoryMismatchTrainTest',
@@ -39,4 +40,5 @@ __all__ = [
     'TrainTestLabelDrift',
     'TrainTestSamplesMix',
     'WholeDatasetDrift',
+    'MultivariateDrift'
 ]
